+++ account-kit/wallet-client/src/client/actions/signSignatureRequest.ts
@@ -1,15 +1,20 @@
 import type { SmartAccountSigner } from "@aa-sdk/core";
-import { type Hex, hexToNumber, serializeSignature } from "viem";
-import { assertNever } from "../../utils.js";
-import {
-  type PersonalSignSignatureRequest,
-  type TypedDataSignatureRequest,
-  type AuthorizationSignatureRequest,
-  type Eip7702UnsignedAuth,
+import { hexToNumber, serializeSignature } from "viem";
+import { assertNever, isWebAuthnSigner } from "../../utils.js";
+import type {
+  PersonalSignSignatureRequest,
+  TypedDataSignatureRequest,
+  AuthorizationSignatureRequest,
+  Eip7702UnsignedAuth,
+  UserOpSig,
+  EcdsaSig,
+  WebauthnSig,
 } from "@alchemy/wallet-api-types";
 import { vToYParity } from "ox/Signature";
 import type { WithoutRawPayload } from "../../types.ts";
 import { metrics } from "../../metrics.js";
+import type { SmartWalletSigner, WebAuthnSigner } from "../index.js";
+import { toWebAuthnAccount } from "viem/account-abstraction";
 
 export type SignSignatureRequestParams = WithoutRawPayload<
   | PersonalSignSignatureRequest
@@ -19,16 +24,41 @@ export type SignSignatureRequestParams = WithoutRawPayload<
     })
 >;
 
-export type SignSignatureRequestResult = {
-  type: "secp256k1";
-  data: Hex;
-};
+export type SignSignatureRequestResult = UserOpSig["signature"];
+
+// Overload: Always an ECDSA signer, can sign any type of request
+export async function signSignatureRequest(
+  signer: SmartAccountSigner,
+  params: WithoutRawPayload<
+    | PersonalSignSignatureRequest
+    | TypedDataSignatureRequest
+    | (AuthorizationSignatureRequest & {
+        data: Eip7702UnsignedAuth;
+      })
+  >,
+): Promise<EcdsaSig["signature"]>;
+
+// Overload: WebAuthn signer, can only sign personal_sign and eth_signTypedData_v4
+export async function signSignatureRequest(
+  signer: WebAuthnSigner,
+  params: WithoutRawPayload<
+    PersonalSignSignatureRequest | TypedDataSignatureRequest
+  >,
+): Promise<WebauthnSig["signature"]>;
+
+// Overload: Union type of signer, can only sign personal_sign and eth_signTypedData_v4
+export async function signSignatureRequest(
+  signer: SmartAccountSigner | WebAuthnSigner,
+  params: WithoutRawPayload<
+    PersonalSignSignatureRequest | TypedDataSignatureRequest
+  >,
+): Promise<UserOpSig["signature"]>;
 
 /**
  * Signs a signature request using the provided signer.
  * This method handles different types of signature requests including personal_sign, eth_signTypedData_v4, and authorization.
  *
- * @param {SmartAccountSigner} signer - The signer to use for signing the request
+ * @param {SmartAccountSigner | WebAuthnSigner} signer - The signer to use for signing the request
  * @param {SignSignatureRequestParams} params - The signature request parameters
  * @param {string} params.type - The type of signature request ('personal_sign', 'eth_signTypedData_v4', or 'signature_with_authorization')
  * @param {SignSignatureRequestParams["data"]} params.data - The data to sign, format depends on the signature type
@@ -56,7 +86,7 @@ export type SignSignatureRequestResult = {
  */
 
 export async function signSignatureRequest(
-  signer: SmartAccountSigner,
+  signer: SmartWalletSigner,
   params: SignSignatureRequestParams,
 ): Promise<SignSignatureRequestResult> {
   metrics.trackEvent({
@@ -68,18 +98,55 @@ export async function signSignatureRequest(
 
   switch (params.type) {
     case "personal_sign": {
+      if (isWebAuthnSigner(signer)) {
+        const webAuthnAccount = toWebAuthnAccount({ ...signer });
+
+        const { signature, webauthn: metadata } =
+          await webAuthnAccount.signMessage({
+            message: params.data,
+          });
+
+        return {
+          type: "webauthn-p256",
+          data: {
+            signature,
+            metadata,
+          },
+        };
+      }
+
       return {
         type: "secp256k1",
         data: await signer.signMessage(params.data),
       };
     }
     case "eth_signTypedData_v4": {
+      if (isWebAuthnSigner(signer)) {
+        const webAuthnAccount = toWebAuthnAccount({ ...signer });
+
+        const { signature, webauthn: metadata } =
+          await webAuthnAccount.signTypedData(params.data);
+
+        return {
+          type: "webauthn-p256",
+          data: {
+            signature,
+            metadata,
+          },
+        };
+      }
       return {
         type: "secp256k1",
         data: await signer.signTypedData(params.data),
       };
     }
     case "eip7702Auth": {
+      if (isWebAuthnSigner(signer)) {
+        throw new Error(
+          "WebAuthn account cannot sign EIP-7702 authorization requests",
+        );
+      }
+
       if (!signer.signAuthorization) {
         throw new Error("Signer does not implement signAuthorization");
       }